feat(Editor): Delete whole word when pressing Mod-Backspace
