+++ packages/components/src/components/editor/editor.tsx
@@ -20,33 +20,10 @@ import { javascript } from '@codemirror/next/lang-javascript'
 import { bracketMatching } from '@codemirror/next/matchbrackets'
 import { multipleSelections } from '@codemirror/next/multiple-selections'
 import { specialChars } from '@codemirror/next/special-chars'
-import { EditorState, SelectionRange } from '@codemirror/next/state'
+import { EditorState } from '@codemirror/next/state'
 import { Command, EditorView } from '@codemirror/next/view'
 import { Component, Element, h, Host, Method, Prop } from '@stencil/core'
-
-/** CodeMirror command to delete content till line start */
-const deleteLine = (view: EditorView) => {
-  const transaction = view.state.t().forEachRange((range, transaction) => {
-    let { from, to } = range
-    if (from === to) {
-      const target = view.movePos(
-        range.head,
-        'backward',
-        'lineboundary',
-        'move'
-      )
-      from = Math.min(from, target)
-      to = Math.max(to, target)
-    }
-    if (from === to) return range
-    transaction.replace(from, to, '')
-    return new SelectionRange(from)
-  })
-  if (!transaction.docChanged) return false
-
-  view.dispatch(transaction.scrollIntoView())
-  return true
-}
+import { deleteLine, deleteWord } from './commands'
 
 export interface EditorContents {
   text: string
@@ -179,11 +156,10 @@ export class Editor {
         'Mod-Enter': this.execute,
         'Mod-ArrowLeft': moveLineStart,
         'Mod-ArrowRight': moveLineEnd,
-        'Ctrl-Space': startCompletion,
+        'Alt-Space': startCompletion,
+        'Alt-Backspace': deleteWord,
         'Mod-Backspace': deleteLine,
-        // FIXME: The following commands have no effect
-        /* 'Mod-Shift-ArrowLeft': selectDocStart, */
-        /* 'Mod-Shift-ArrowRight': selectDocEnd, */
+        // FIXME: Add indentation commands
         ...this.keymap,
       }),
       keymap(baseKeymap),
